feat: add card state types
